// src/app/Admin/Models/iuser-data.ts
export interface IUserData {
    id: number
  firstName?: string
  lastName: string
  email: string
  phoneNumber: string
  dateOfBirth: string
  gender: string
  profilePictureUrl: string
  accountStatus: string
  emailVerified: boolean
  phoneVerified: boolean
  createdAt: string
  updatedAt: string
  lastLogin: string
  role: string
}